Major code change
Now if you want all the details of a meme in a standard array, you can
use meme(). This gets called in memeFeed() and profile() and all future
ones.

Yes. This is inefficient to create a new database connection every time
you go through a list like that. But. It means that it’s much easier to
manage the code, like if you want to add more features to the standard
meme array that is outputted.

Also, now the follow button shows up everywhere (it’s in displayMeme())
and works.