Share and comment links
You can now share the meme with the share action provider and you can
tap on a users face/name in the comments to go to their profile.
Unfortunately the whole thing has to get rebuilt when you go back to it
because the view gets squished for no reason - no easy solution apart
from switch platforms.